services/identity/requirements.txt
# Identity Service Dependencies
Django>=5.2
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
django-filter>=25.2
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
whitenoise>=6.11.0
gunicorn>=23.0.0
msal>=1.34.0
PyJWT[crypto]>=2.10.1
cryptography>=46.0.3
redis>=5.0.0
requests>=2.31.0
